refactor: Simplify blake3 fuzzing
Create a method for pushing blake3 input, for computing the hash of that
input, and for parsing the resulting output. Use these functions in the
blake3 fuzz test.

I drop the parameter that toggles full / half tables. The caller likely
doesn't understand the difference, nor should he. Always using full
tables is a sane default and removing a function parameter simplifies
the API. Testing both half and full tables adds no value and merely
makes the tests slower.

In the following commits, I will port the remaining unit tests to the
new API.