refactor: Blake3 fuzzing
Run a cached version of the blake3 bitcoin script on many different
inputs. Because blake3 has a different script for each byte length,
I fixed the input to 32 bytes.

Reviewers might point out that the fuzz test is pointless if it doesn't
check all possible byte lengths. I argue that we are already testing the
minimal and maximal length, as well as all single bytes, in unit tests.
Generating blake3 scripts for multiple lengths takes considerable time
when fuzzing is started (even for 0 to 17 bytes, the setup takes ~10
seconds). This is why I argue, it is more effective to explore the space
of 32-byte inputs to a deep level, instead of exploring the space of 0
to 1024-byte inputs on a shallow level.