refactor: Make blake3 private
Use `blake3_compute_script` as the main function for generating BLAKE3
Bitcoin scripts, in order to hide implementation details (define_var,
use_full_tables). Move the documentation over and restructure it to make
it clearer. Update the only call to `blake3` inside the chunker. Add a
safety comment about u32 and usize. (In Rust, it is easier to keep lengths
of things as type usize than to keep them as type u32.)

# Conflicts:
#	bitvm/src/chunk/wrap_hasher.rs