Refactor blake3 (#257)
* refactor: Remove useless parameter

The function was always called with constants = vec![],
so the constants had no effect at all.

* refactor: Remove useless parameter

The function was always called with nibble_count = 8,
so we can inline the parameter.

* feat: Execute ScriptBuf

There should exist a function that lets you execute a byte string of
Bitcoin Script, aka ScriptBuf. This is especially useful when a script
is constructed by concatenating multiple pieces of precompiled code.

* refactor: Simplify blake3 fuzzing

Create a method for pushing blake3 input, for computing the hash of that
input, and for parsing the resulting output. Use these functions in the
blake3 fuzz test.

I drop the parameter that toggles full / half tables. The caller likely
doesn't understand the difference, nor should he. Always using full
tables is a sane default and removing a function parameter simplifies
the API. Testing both half and full tables adds no value and merely
makes the tests slower.

In the following commits, I will port the remaining unit tests to the
new API.

* refactor: Clean up blake3 unit tests

Use the predefined functions for input / computation / output.
Remove the randomized unit tests. These are superseded by fuzzing.
Replace the tests "all zeroes" / "all ones" with a test that hashes
every single byte.

* refactor: Blake3 fuzzing

Run a cached version of the blake3 bitcoin script on many different
inputs. Because blake3 has a different script for each byte length,
I fixed the input to 32 bytes.

Reviewers might point out that the fuzz test is pointless if it doesn't
check all possible byte lengths. I argue that we are already testing the
minimal and maximal length, as well as all single bytes, in unit tests.
Generating blake3 scripts for multiple lengths takes considerable time
when fuzzing is started (even for 0 to 17 bytes, the setup takes ~10
seconds). This is why I argue, it is more effective to explore the space
of 32-byte inputs to a deep level, instead of exploring the space of 0
to 1024-byte inputs on a shallow level.

* refactor: Blake3 fuzz target

Cover all message sizes that are used inside BitVM.

* refactor: Make blake3 private

Use `blake3_compute_script` as the main function for generating BLAKE3
Bitcoin scripts, in order to hide implementation details (define_var,
use_full_tables). Move the documentation over and restructure it to make
it clearer. Update the only call to `blake3` inside the chunker. Add a
safety comment about u32 and usize. (In Rust, it is easier to keep lengths
of things as type usize than to keep them as type u32.)

# Conflicts:
#	bitvm/src/chunk/wrap_hasher.rs