refactor: Rename "blocks" to "digits"
This commit renames "blocks" to "digits" in function signatures and
struct members. I used the opportunity to clearly separate lengths
in units of digits from lengths in units of bits.

This commit touches a lot of code. The Parameter struct is technically
public and changed by this commit. However, I argue that no one should
manually set up his parameters. Instead, everyone should use the
high-level methods for wots-hash, wots-blake and wots256, which are not
affected by this commit.

I plan to make the Parameter struct private in a follow-up PR.