WOTS data format refactor (#289)
* chore: remove paste macro

I simply transform `paste! { X }` into `X`. If the diff for this commit
is large, then git is having an issue. The `paste!` macro is used
elsewhere in the codebase, so I won't remove the dependency.

* chore: clippy

cargo clippy --fix --lib -p bitvm

* chore: replace static with const

* chore: Inline

* refactor: Replace macros with functions

Functions are easier to read and tools like `rustfmt` and `clippy` have
an easier time parsing them.

* refactor: clarify endianness

* doc: winternitz

* refactor: Rename "blocks" to "digits"

This commit renames "blocks" to "digits" in function signatures and
struct members. I used the opportunity to clearly separate lengths
in units of digits from lengths in units of bits.

This commit touches a lot of code. The Parameter struct is technically
public and changed by this commit. However, I argue that no one should
manually set up his parameters. Instead, everyone should use the
high-level methods for wots-hash, wots-blake and wots256, which are not
affected by this commit.

I plan to make the Parameter struct private in a follow-up PR.

* doc: address comments

Only changes to docstrings and parameter names.
No changes to public API.

* fix: documentation