refactor: adapt bridge to new wots api
Removes the dependency on wots_api.
Removing signing_winternitz turned out to be too much work, as it
touches a lot of code. In particular, the connector outputs handle
public keys for different message lengths in the same vector.
The bridge code can be refactored in the future if needed.