Updated WOTS API (#293)
* chore: Mark potential footgun

* feat: New WOTS API

* test: Add WOTS test vectors

The test vectors are generated directly in Rust, so we can add new
vectors or modify existing vectors in the future.

* refactor: Adapt bitvm crate to new WOTS API

Removes the dependency on the modules wots_api and signing_winternitz.

* refactor: adapt bridge to new wots api

Removes the dependency on wots_api.
Removing signing_winternitz turned out to be too much work, as it
touches a lot of code. In particular, the connector outputs handle
public keys for different message lengths in the same vector.
The bridge code can be refactored in the future if needed.

* refactor: Move winternitz_hash into signing_winternitz

signing_winternitz is the only place that calls winternitz_hash,
so why not combine the modules?

BLAKE3 switched from 20-byte WOTS signatures to Wots16 a while ago,
so I use the opportunity to remove the code for 20-byte WOTS.

* chore: Remove wots_api.rs

No more code depends on this, so it is safe to remove.

* fix: Address comments

* refactor: WinternitzSecret constructors

Deprecate from_string constructors. Add from_bytes as a safer
alternative.

* fix: Address comments

* fix: address comments

* fix: cargo fmt